Send debug logs to files + numerous other updates

+++ scripts/adbalerts.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
+# adbalerts.sh
+# 2025.05.05
 
+script=$(basename "$0" | sed 's/\.sh$//')
+exec 3> /config/$script.debug.log
+BASH_XTRACEFD=3
 set -x
 
 dvr="$1"
@@ -8,14 +13,16 @@ channelsPort=$(echo "$dvr" | awk -F: '{print $2}')
 foregroundScript=adbalerts
 runningScriptPID=$(ps -ef | grep "[a]dbalerter.sh $dvr" | awk '{print $2}')
 greenIcon=\"icons\/channels.png\"
+#greenIcon=\"custom-webui\/icons\/channels.png\"
 purpleIcon=\"https:\/\/community-assets.getchannels.com\/original/2X/5/55232547f7e8f243069080b6aec0c71872f0f537.png\"
 logFile=/config/"$channelsHost"-"$channelsPort"_"$foregroundScript"_latest.log
-  rm $logFile
+  [[ -f $logFile && $PERSISTENT_LOGS != "true" ]] && rm $logFile
 configFile=/config/config.yaml
 configTemp=/tmp/config.yaml
 
 #Trap end of script run
 finish() {
+  #nohup /config/finish.sh $configTemp >> $logFile 2>&1 &
   cp $configTemp /config
 }
